Updates tests & matrix, switching to testing GitLab EE
Fixes broken EE tests. It looks like the license file was trying to write to a
directory called `license/`, but was conflicting with the file named `LICENSE`.

This switches the default tests to use GitLab EE rather than CE. The two
installations are the same thing, but the presence of the license file enables
additional features.

From the GitLab info page:

"This means that if you have installed GitLab Enterprise Edition without a
license, you will not notice a difference with a typical Community Edition
instance [...]"