Debug to detect when `default_branch` is set in .tf file
WIP for fixing #299 where Terraform tries to reset the
`default_branch` to null after a repository is added on
GitLab side.

The `default_branch` defaults to `master` on GitLab side,
but it is only set if the repository is not empty.
Otherwise it is `null` on remote side.

The commit generates the following output, which shows
no trace of `null` value from remote.

    2020/06/19 20:55:55 [DEBUG] gitlab.Project default_branch (string) ""
    2020/06/19 20:55:55 [DEBUG] d.Get default_branch (string) ""
    2020/06/19 20:55:55 [DEBUG] d.GetOk default_branch (string) "" - false
    2020/06/19 20:55:55 [DEBUG] d.HasChange default_branch false
    2020/06/19 20:55:55 [DEBUG] d.GetChange default_branch (string) "", (string) ""

But the actual response shows that `null` was returned in
JSON.

    2020/06/19 20:55:55 [DEBUG] GitLab API Response Details:
    ...
     "default_branch": null,