gitlab_project_mirror: Fix url attribute drift when using basicauth
The fix is to conditionally set the "url" attribute only if it does not contain redacted text.

The caveat is that the "url" attribute cannot be imported if it has redacted text. I don't expect this to be a problem. The attribute is `ForceNew: true` so any upstream change to the value is not possible.

As a bonus I fixed the destroy test to actually test the delete function, by creating the Gitlab project outside of the test state. I moved `testAccGitlabProjectContext` into `helper_test.go` now that it is used more than once, but didn't modify it.