Support `early_auth_check` flag in provider config. Closes #777
This change adds a new `early_auth_check` flag to the provider configuration.
It defaults to `true` which causes the provider to perform a dummy request to the GitLab API to verify the
configuration. If that is not desired it can be set to `false` which avoids this call.
This can be in useful if the GitLab instance is created within the same terraform module as this provider
is being used.