Implement `archive_on_destroy` attribute for projects. Closes #761
This change implements an `archive_on_destroy` attribute on the
`gitlab_project` resource. It defaults to `false`, if set to `true`, on
a destroy it will *archive* the project instead of deleting it.

This is especially useful in organizations where repositories need to
be kept for regulatory purposes and can't be deleted.