resource/service_slack: use pass through importer and start alignment with upstream API. Refs #790
This is a start towards correctly implementing the Slack Notification
Integration Service. When I started I wasn't aware of the problems in
the upstream GitLab API and the misalignment between `go-gitlab` and the
upstream GitLab API (maybe by intention because the upstream API is broken).

Long story short, there is a GitLab upstream issue here:

* https://gitlab.com/gitlab-org/gitlab/-/issues/28903

a go-gitlab issue here:

* https://github.com/xanzy/go-gitlab/issues/1354

And we have one here:

* https://github.com/gitlabhq/terraform-provider-gitlab/issues/790

Nevertheless, I think we should merge this change set. It uses the pass
through importer for the resource and also cleans up a few things. In
addition I've aligned the resource schema with the upstream API and
commented out the missing parts with matching comments and links to the
issues.
I'd say we continue work on this once both GitLab and go-gitlab APIs are
fixed / aligned.