Add managed license resources and tests (#700)
* Add managed license resources and tests

* Add documentation for the managed_license resource

* Update to v2 schema using context

This involves several updates: 1. Update the schema.Resource to the "Context" method; 2. Pass in context to each of the functions and update returns to "diagnostic" type; 3. Pass in context to each of the client calls.

* Add import statement and update tests accordingly

In addition, I added the import statement the documentation and renamed the documentation filename to remove the "gitlab_" from the front which now matches the convention of the remainder of the documentation.

* Pass through updating documentation to new standards and using new SkipFunc

* Fixed several issues with tests. Implemented specific importer methods.

The importer passthrough method seemed to break on the multi-ID needed for the project:license pair.

* Rename license variable from "id" to "licenseId"

* Update import to use StatePassthroughContext

This also required updating the CRUD methods to accept a 2 part string and parse out the ID

* Remove dead code.

* Remove random license name

* Apply suggestion re: removing from state if it doesn't exist in upsteam

Co-authored-by: Adam Snyder <armsnyder@gmail.com>

* Add callout re: Ultimate subscription and comments to the import.sh

* Properly generate documentation for pr-lint.

* Added second newline and re-generated documentation.

Co-authored-by: Adam Snyder <armsnyder@gmail.com>