resource/gitlab_project_approval_rule: fix potential accidential destroy of resource if not in first page
This fixes a potential bug that the resource is marked for destroy after a read,
because it was not found on the first page. Pagination was not implemented.
However, there exist a specific API to get a single project-level approval rule.
I have implemented it in go-gitlab:

* https://github.com/xanzy/go-gitlab/pull/1410

After it is released, we should be able to just rebase and LGTM ;)