Fix a case where a change to a project in terraform can never apply when certain fields are modified.
Because the gitlab API actually expects at least on field out of a subset of fields in an update call (aside from `id`), if you only modify fields not in this subset, the Gitlab API will return a 400 complaining that one of the needed fields is missing.

One of these special fields is `name`. As this seems to be a safe field to assume always exists (we'll still defensively check if nil though), we pass it in always.