resource/gitlab_application_settings: New experimental resource
This change set implements the new `gitlab_application_settings`
resource which provides a way to change the GitLab Application Settings
using the API at
https://docs.gitlab.com/ee/api/settings.html#change-application-settings.

Even though I suggest that we can merge this first experimental version
of the resource there are still quite some open questions for me.

I'd like to tackle them over at
https://github.com/gitlabhq/terraform-provider-gitlab/issues/957 and
give the community the change to take part in the discussion on how to
move forward with this.

Refs: #957 #1187